refactor: set_heredoc_utils.cのファイル分割、set_redirectionのinputとoutputの関数を切り出す
